src/slash/ 以下のリファクタリング
